Move the task form to its own component
